requirements.txt
packaging=24.2
pyyaml=6.0.2